requirements/main.txt
celery==5.1.0
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
factory-boy==2.12.0
psycopg2==2.8.6
asana==0.10.0
djangorestframework==3.12.4
